+++ SQLiteStudio3/coreSQLiteStudio/plugins/scriptingsql.cpp
@@ -38,7 +38,7 @@ void ScriptingSql::resetContext(ScriptingPlugin::Context* context)
     dynamic_cast<SqlContext*>(context)->errorText.clear();
 }
 
-QVariant ScriptingSql::evaluate(ScriptingPlugin::Context* context, const QString& code, const QList<QVariant>& args, Db* db, bool locking)
+QVariant ScriptingSql::evaluate(ScriptingPlugin::Context* context, const QString& code, const FunctionInfo& funcInfo, const QList<QVariant>& args, Db* db, bool locking)
 {
     SqlContext* ctx = dynamic_cast<SqlContext*>(context);
     ctx->errorText.clear();
@@ -58,13 +58,14 @@ QVariant ScriptingSql::evaluate(ScriptingPlugin::Context* context, const QString
     QString sql = code;
     if (ctx->variables.size() > 0)
     {
-        QString value;
         for (const QString& key : ctx->variables.keys())
         {
-            value = "'" + ctx->variables[key].toString() + "'";
+            QString value = "'" + ctx->variables[key].toString() + "'";
             sql.replace(":" + key, value).replace("@" + key, value).replace("$" + key, value);
         }
     }
+
+    replaceNamedArgs(sql, funcInfo, args);
 
     SqlQueryPtr result = theDb->exec(sql, args, execFlags);
     if (result->isError())
@@ -76,7 +77,7 @@ QVariant ScriptingSql::evaluate(ScriptingPlugin::Context* context, const QString
     return result->getSingleCell();
 }
 
-QVariant ScriptingSql::evaluate(const QString& code, const QList<QVariant>& args, Db* db, bool locking, QString* errorMessage)
+QVariant ScriptingSql::evaluate(const QString& code, const FunctionInfo& funcInfo, const QList<QVariant>& args, Db* db, bool locking, QString* errorMessage)
 {
     Db* theDb = nullptr;
 
@@ -91,7 +92,10 @@ QVariant ScriptingSql::evaluate(const QString& code, const QList<QVariant>& args
     if (!locking)
         execFlags |= Db::Flag::NO_LOCK;
 
-    SqlQueryPtr result = theDb->exec(code, args, execFlags);
+    QString sql = code;
+    replaceNamedArgs(sql, funcInfo, args);
+
+    SqlQueryPtr result = theDb->exec(sql, args, execFlags);
     if (result->isError())
     {
         *errorMessage = result->getErrorText();
@@ -143,4 +147,19 @@ void ScriptingSql::deinit()
     contexts.clear();
 
     safe_delete(memDb);
+}
+
+void ScriptingSql::replaceNamedArgs(QString& sql, const ScriptingPlugin::FunctionInfo& funcInfo, const QList<QVariant>& args)
+{
+    int i = 0;
+    for (const QString& key : funcInfo.getArguments())
+    {
+        if (i >= args.size())
+            break;
+
+        QString value = "'" + args[i++].toString() + "'";
+        sql.replace(":" + key, value)
+           .replace("@" + key, value)
+           .replace("$" + key, value);
+    }
 }